Refactor: moved databaseManager into database package, also move for test as well
