+++ Simple Banking System (Java)/task/src/banking/database/DatabaseManager.java
@@ -1,4 +1,4 @@
-package banking;
+package banking.database;
 
 
 import java.sql.Connection;
