+++ Simple Banking System (Java)/task/test/banking/database/DatabaseManagerTest.java
@@ -1,4 +1,4 @@
-package banking;
+package banking.database;
 
 import org.junit.Test;
